renamed ApplicationLite to AppBase
